Stage 2 of fixing psi dependence on input electron positions
Now I understand how 95% of the network works.
Updated some of the code.
Tons of documentation/examples using methane as a reference to better understand the dimensionality.
With current changes network runs fine but psi still does not have dependence
on input x values.

My next guess is to look at the orbital calculation?
or another option is to look at the computational graph?
Should we just replace the matmuls with convolutions?? as per ferminet?